+++ flake.nix
@@ -1,84 +1,62 @@
 {
-  description = "Cicero jobs populating https://cache.iog.io – this cannot yet build Daedalus";
+  description = "The open source wallet for ada, built to grow with the Cardano blockchain";
+
   inputs = {
-    nixpkgs.follows = "tullia/nixpkgs";
+    nixpkgs.url = "github:nixos/nixpkgs/nixpkgs-22.11-darwin";
+    cardano-wallet-unpatched.url = "github:input-output-hk/cardano-wallet/v2023-04-14";
+    cardano-wallet-unpatched.flake = false; # otherwise, +10k quadratic dependencies in flake.lock…
+    cardano-world.url = "github:input-output-hk/cardano-world/a0a315100ee320395be97fcc83f46678d5a7fb6e";
+    cardano-world.flake = false; # otherwise, +19k quadratic dependencies in flake.lock…
+    cardano-shell.url = "github:input-output-hk/cardano-shell/0d1d5f036c73d18e641412d2c58d4acda592d493";
+    cardano-shell.flake = false;
     tullia.url = "github:input-output-hk/tullia";
+    tullia.flake = false; # otherwie, +1k dependencies in flake.lock…
+    flake-compat.url = "github:input-output-hk/flake-compat";
+    flake-compat.flake = false;
   };
-  outputs = inputs: {
-    hydraJobs = rec {
 
-      # --- Linux ----------------------------------------------------
-      x86_64-linux.x86_64-linux = let
-        d = import ./default.nix { target = "x86_64-linux"; localLibSystem = "x86_64-linux"; };
-      in {
-        cardano-bridge = d.daedalus-bridge;
-        cardano-node = d.cardano-node;
-        # daedalus = d.daedalus;  # TODO: I’m really not sure if it still makes sense, if we have Buildkite…
-        daedalus-installer = d.daedalus-installer;
-        devShellGCRoot = (import ./shell.nix { system = "x86_64-linux"; autoStartBackend = true; }).gcRoot;
-        mono = d.pkgs.mono;
-        nodejs = d.nodejs;
-        tests = d.tests;
-        wine = d.wine;
-        wine64 = d.wine64;
-        yaml2json = d.yaml2json;
-      };
-      # --------------------------------------------------------------
-
-      # --- Windows (x-compiled from Linux) --------------------------
-      x86_64-linux.x86_64-windows = let
-        d = import ./default.nix { target = "x86_64-windows"; localLibSystem = "x86_64-linux"; };
-      in {
-        cardano-bridge = d.daedalus-bridge;
-        cardano-node = d.cardano-node;
-      };
-      # --------------------------------------------------------------
-
-      # --- Darwin ---------------------------------------------------
-      x86_64-darwin.x86_64-darwin = let
-        d = import ./default.nix { target = "x86_64-darwin"; localLibSystem = "x86_64-darwin"; };
-      in {
-        cardano-bridge = d.daedalus-bridge;
-        cardano-node = d.cardano-node;
-        daedalus-installer = d.daedalus-installer;
-        devShellGCRoot = (import ./shell.nix { system = "x86_64-darwin"; autoStartBackend = true; }).gcRoot;
-        nodejs = d.nodejs;
-        yaml2json = d.yaml2json;
-      };
-      # --------------------------------------------------------------
-
-      # --- What CI should build -------------------------------------
-      x86_64-linux.required = inputs.nixpkgs.legacyPackages.x86_64-linux.releaseTools.aggregate {
-        name = "required for CI";
-        constituents = __attrValues x86_64-linux.x86_64-linux ++ __attrValues x86_64-linux.x86_64-windows;
-      };
-      x86_64-darwin.required = inputs.nixpkgs.legacyPackages.x86_64-darwin.releaseTools.aggregate {
-        name = "required for CI";
-        constituents = __attrValues x86_64-darwin.x86_64-darwin;
-      };
-      # --------------------------------------------------------------
-
-    };
-  }
-  // (let
-    x86_64-linux  = inputs.tullia.fromSimple "x86_64-linux"  (import ./nix/tullia.nix);
-    x86_64-darwin = inputs.tullia.fromSimple "x86_64-darwin" (import ./nix/tullia.nix);
-    fakeEvent = { inputs."GitHub event" = {id = ""; created_at = ""; value = {github_body.head_commit.id="0000000";};}; id=""; ociRegistry=""; };
+  outputs = inputs: let
+    supportedSystems = ["x86_64-linux" "x86_64-darwin" "aarch64-darwin"];
+    inherit (inputs.nixpkgs) lib;
   in {
-    tullia.x86_64-linux = x86_64-linux.tullia;
-    cicero.x86_64-linux = x86_64-linux.cicero;
-    tullia.x86_64-darwin = x86_64-darwin.tullia;
-    cicero.x86_64-darwin = x86_64-darwin.cicero;
+    internal = import ./nix/internal.nix { inherit inputs; };
 
-    ciceroLocalTest.x86_64-linux  = (x86_64-linux.cicero."daedalus/ci"  fakeEvent).job;
-    ciceroLocalTest.x86_64-darwin = (x86_64-darwin.cicero."daedalus/ci" fakeEvent).job;
+    packages = lib.genAttrs supportedSystems (buildSystem:
+      import ./nix/packages.nix { inherit inputs buildSystem; }
+    );
 
-  });
-  # --- Flake Local Nix Configuration ----------------------------
+    devShells = lib.genAttrs supportedSystems (
+      system: let
+        all = lib.genAttrs inputs.self.internal.installerClusters (
+          cluster: import ./nix/internal/old-shell.nix { inherit inputs system cluster; }
+        );
+      in all // { default = all.mainnet; }
+    );
+
+    # Compatibility with older Nix:
+    defaultPackage = __mapAttrs (_: a: a.default) inputs.self.outputs.packages;
+    devShell = __mapAttrs (_: a: a.default) inputs.self.outputs.devShells;
+
+    hydraJobs = {
+      installer = lib.genAttrs (supportedSystems ++ ["x86_64-windows"]) (
+        targetSystem: lib.genAttrs inputs.self.internal.installerClusters (
+          cluster: inputs.self.internal.${targetSystem}.${cluster}.unsignedInstaller
+        )
+      );
+      devshell = lib.genAttrs supportedSystems (system: inputs.self.devShells.${system}.default);
+      required = inputs.nixpkgs.legacyPackages.x86_64-linux.releaseTools.aggregate {
+        name = "github-required";
+        meta.description = "All jobs required to pass CI";
+        constituents =
+          lib.collect lib.isDerivation inputs.self.hydraJobs.installer
+          ++ lib.collect lib.isDerivation inputs.self.hydraJobs.devshell;
+      };
+    };
+  };
+
   nixConfig = {
-    extra-substituters = ["https://cache.iog.io" "https://iog.cachix.org"];
-    extra-trusted-public-keys = ["hydra.iohk.io:f/Ea+s+dFdN+3Y/G+FDgSq+a5NEWhJGzdjvKNGv0/EQ=" "iog.cachix.org-1:nYO0M9xTk/s5t1Bs9asZ/Sww/1Kt/hRhkLP0Hhv/ctY="];
+    extra-substituters = ["https://cache.iog.io"];
+    extra-trusted-public-keys = ["hydra.iohk.io:f/Ea+s+dFdN+3Y/G+FDgSq+a5NEWhJGzdjvKNGv0/EQ="];
     allow-import-from-derivation = "true";
   };
-  # --------------------------------------------------------------
 }