+++ source/renderer/app/api/voting/requests/getCatalystFund.ts
@@ -1,11 +1,17 @@
 import { externalRequest } from '../../utils/externalRequest';
-import { MAINNET_SERVICING_STATION_URL } from '../../../config/urlsConfig';
+import { CATALYST_API_URL } from '../../../config/urlsConfig';
 import { GetCatalystFundResponse } from '../types';
 
-export const getCatalystFund = (): Promise<GetCatalystFundResponse> =>
-  externalRequest({
-    hostname: MAINNET_SERVICING_STATION_URL,
+export const getCatalystFund = (): Promise<GetCatalystFundResponse> => {
+  const urlOverride: URL | undefined = environment.catalystApiUrlOverride
+    ? new URL(environment.catalystApiUrlOverride)
+    : undefined;
+
+  return externalRequest({
+    hostname: urlOverride ? urlOverride.hostname : CATALYST_API_URL,
     path: '/api/v0/fund',
     method: 'GET',
-    protocol: 'https',
+    port: urlOverride ? Number(urlOverride.port) : undefined,
+    protocol: urlOverride ? urlOverride.protocol.replace(':', '') : 'https',
   });
+};