+++ source/renderer/app/components/voting/voting-info/RegisterToVote.tsx
@@ -2,64 +2,32 @@ import React, { useState } from 'react';
 import { injectIntl } from 'react-intl';
 import { Button } from 'react-polymorph/lib/components/Button';
 import { Checkbox } from 'react-polymorph/lib/components/Checkbox';
-import {
-  formattedDateTime,
-  mapToLongDateTimeFormat,
-} from '../../../utils/formatters';
-import type { Locale } from '../../../../../common/types/locales.types';
 import type { Intl } from '../../../types/i18nTypes';
 import { messages } from './RegisterToVote.messages';
 import { messages as votingMessages } from './VotingInfo.messages';
 import styles from './RegisterToVote.scss';
-import votingStyles from './VotingInfo.scss';
-import type { CatalystFund } from '../../../api/voting/types';
 
 type Props = {
-  currentLocale: Locale;
-  currentDateFormat: string;
-  currentTimeFormat: string;
-  fundInfo: CatalystFund;
   intl: Intl;
   onRegisterToVoteClick: (...args: Array<any>) => any;
 };
 
-function RegisterToVote({
-  currentLocale,
-  currentDateFormat,
-  currentTimeFormat,
-  fundInfo,
-  intl,
-  onRegisterToVoteClick,
-}: Props) {
+function RegisterToVote({ intl, onRegisterToVoteClick }: Props) {
   const [step1, setStep1] = useState(false);
   const [step2, setStep2] = useState(false);
   const canRegister = step1 && step2;
-  const snapshotDate = formattedDateTime(
-    fundInfo.next.registrationSnapshotTime,
-    {
-      currentLocale,
-      ...mapToLongDateTimeFormat({
-        currentLocale,
-        currentDateFormat,
-        currentTimeFormat,
-      }),
-    }
-  );
+
   return (
     <div className={styles.root}>
-      <span className={styles.title}>
-        {intl.formatMessage(votingMessages.fundName, {
-          votingFundNumber: fundInfo.next.number,
-        })}
-      </span>
-      <span className={styles.dateLabel}>
-        {intl.formatMessage(messages.dateLabel)}
-      </span>
-      <span className={styles.date}>{snapshotDate}</span>
-      <hr className={votingStyles.separator} />
-      <span className={styles.stepsTitle}>
+      <div className={styles.title}>
+        {intl.formatMessage(votingMessages.registerToVoteHeader)}
+      </div>
+      <div className={styles.votingInstructions}>
+        {intl.formatMessage(messages.votingInstructions)}
+      </div>
+      <div className={styles.stepsTitle}>
         {intl.formatMessage(messages.stepsTitle)}
-      </span>
+      </div>
       <div className={styles.step}>
         <Checkbox
           checked={step1}